add modal

+++ client/src/app/components/room/roomList/roomList.component.ts
@@ -1,6 +1,8 @@
 import { Component } from "@angular/core";
 import { Router } from "@angular/router"
-import { MdSnackBar } from "@angular/material";
+import { MdSnackBar,MdDialog } from "@angular/material";
+import {ConfirmationModalComponent} from "../../modals/confirmationModal/confirmationModal.component";
+
 import { RoomDataSource } from "../../../dataSources/roomDataSource";
 import { RoomApiService } from "../../../services/api/roomApiService";
 import { Room } from "../../../../../../common/models/room";
@@ -17,7 +19,8 @@ export class RoomListComponent {
 
     constructor(private router: Router,
         private roomApiService: RoomApiService,
-        private snackbar: MdSnackBar) {
+        private snackbar: MdSnackBar,
+        private dialog: MdDialog) {
 
     }
 
@@ -30,16 +33,21 @@ export class RoomListComponent {
     }
 
     deleteRoom(selectedRoom: Room) {
-        this.roomApiService.delete(selectedRoom._id)
-            .then(() => {
-                this.snackbar.open("Salle supprimée avec succès", "X", {
-                    duration: 5000
-                });
-            }).catch((error) => {
-                this.snackbar.open("Erreur lors de la suppression de la salle: " + error, "X", {
-                    duration: 5000
-                });
-            })
+        let modal = this.dialog.open(ConfirmationModalComponent);
+        modal.afterClosed().subscribe((result) => {
+            if(result === "Yes") {
+                this.roomApiService.delete(selectedRoom._id)
+                    .then(() => {
+                        this.snackbar.open("Salle supprimée avec succès", "X", {
+                            duration: 5000
+                        });
+                    }).catch((error) => {
+                        this.snackbar.open("Erreur lors de la suppression de la salle: " + error, "X", {
+                            duration: 5000
+                        });
+                    });
+            }
+        });
     }
 
     deleteRooms() {
@@ -50,6 +58,10 @@ export class RoomListComponent {
 
     ngOnInit() {
         this.listSource = new RoomDataSource();
+    }
+
+    selectRow() {
+        
     }
 
 }
